backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
PyJWT==2.8.0
bcrypt==4.1.2
pyotp==2.9.0

# Google Calendar Integration
google-api-python-client==2.108.0

# AI/LLM Integration
groq==0.4.1
openai==1.3.7

# HTTP Requests and Utilities
httpx==0.25.2
requests==2.31.0

# Async support
aiofiles==23.2.1

# Environment management
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0
dnspython==2.4.2 